adding files module.js
