+++ index.html
@@ -29,7 +29,7 @@
     </div>
   </div>
 
-  <script type="module" src="main.js"></script>
+  <script type="module" src="./module/main.js"></script>
 </body>
 
 </html>
