+++ module/api.js
@@ -1,5 +1,5 @@
-const textElementsLoad = document.querySelector(".text-load");
-textElementsLoad.style.display = "block"
+import { textElementsLoad } from "./main.js";
+
 
 export function getComments() {
     return fetch("https://wedev-api.sky.pro/api/v1/Aleksey-Rudnev/comments", {
@@ -13,4 +13,28 @@ export function getComments() {
             return response
         })
 }
+
+
+
+export const addTodo = (name, text) => {
+    return fetch("https://wedev-api.sky.pro/api/v1/Aleksey-Rudnev/comments", {
+
+        method: "POST",
+        body: JSON.stringify({
+            text: text
+                .replaceAll("&", "&amp;")
+                .replaceAll("<", "&lt;")
+                .replaceAll(">", "&gt;")
+                .replaceAll('"', "&quot;"),
+            name: name
+                .replaceAll("&", "&amp;")
+                .replaceAll("<", "&lt;")
+                .replaceAll(">", "&gt;")
+                .replaceAll('"', "&quot;"),
+            forceError: true
+        })
+    })
+}
+
+
